feat: dispatch work, simplify
